116: Add UserFixturesContext
In order to keep user details across different contexts easily managable
we should keep them in one place and add the context to each suite
