+++ src/User/Behat/UserDomainContext.php
@@ -116,35 +116,4 @@ class UserDomainContext implements Context
         Assert::notNull($this->exception);
         Assert::isInstanceOf($this->exception, \DomainException::class);
     }
-
-    /**
-     * @Transform
-     */
-    public function createUser(string $name): UserEntity
-    {
-        $city = \Mockery::mock(CityEntity::class);
-
-        switch ($name) {
-            case 'Alex Smith':
-                return new UserEntity(
-                    UserId::create(),
-                    'alex.smith',
-                    'Alex',
-                    'Smith',
-                    'alex@example.com',
-                    'passw0rd',
-                    $city
-                );
-            case 'Jo Johnson':
-                return new UserEntity(
-                    UserId::create(),
-                    'jo.johnson',
-                    'Jo',
-                    'Johnson',
-                    'jo@example.com',
-                    'passw0rd',
-                    $city
-                );
-        }
-    }
 }
